requirements.txt
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
blinker==1.9.0
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
deprecation==2.1.0
Flask==3.1.2
Flask-Compress==1.23
gotrue==1.3.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pillow==12.0.0
postgrest==0.10.8
pyaes==1.6.1
pydantic==2.12.5
pydantic_core==2.41.5
Pyrogram==2.0.106
TgCrypto
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
realtime==1.0.6
requests==2.32.5
six==1.17.0
sniffio==1.3.1
storage3==0.5.4
StrEnum==0.4.15
supabase==1.0.4
supafunc==0.2.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
websockets==12.0
Werkzeug==3.1.4
gunicorn
gevent
Flask-WTF>=1.2.0
bleach>=6.0.0
nest_asyncio>=1.6.0
